crypto-bigint: factor `uint` tests into a `tests/` module
This allows using `std`-dependent features in tests, and also would make
it easier to refactor these tests into separate modules based on given
operations or other functionality.
